refactor: move cell type and supporting functions to cell.ts
remove cellSiblings function from game_reducer and refactor code to use 'siblingsOf' cell function
