+++ web-app/src/components/game/game_reducer.ts
@@ -1,4 +1,4 @@
-import { Cell } from "@/shared/common";
+import { Cell, siblingsOf } from "@/shared/cell";
 import { Digit, isDigit } from "@/shared/digit";
 import { Puzzle } from "@/shared/puzzle";
 import { Stack } from "@/shared/stack";
@@ -58,43 +58,13 @@ function setup(puzzle: Puzzle) {
   };
 }
 
-function cellSiblings(index?: number) {
-  if (index === undefined) return new Set<number>();
-
-  const column = index % 9;
-  const row = Math.floor(index / 9);
-  const square = Math.floor(index / 27) * 3 + Math.floor((index % 9) / 3);
-
-  const columnSiblings = Array.from({ length: 9 }).map(
-    (_, i) => column + i * 9,
-  );
-  const rowSiblings = Array.from({ length: 9 }).map((_, i) => row * 9 + i);
-  const squareSiblings = Array.from({ length: 9 }).map(
-    (_, i) =>
-      Math.floor(square / 3) * 27 +
-      (square % 3) * 3 +
-      (i % 3) +
-      Math.floor(i / 3) * 9,
-  );
-
-  const siblingIndices = new Set([
-    ...columnSiblings,
-    ...rowSiblings,
-    ...squareSiblings,
-  ]);
-
-  siblingIndices.delete(index);
-
-  return siblingIndices;
-}
-
 function updateAndPrune(
   state: GameState,
   index: number,
   digit: Digit,
 ): GameState {
   // 1. check for sibling clash
-  const siblings = cellSiblings(index);
+  const siblings = siblingsOf(index);
 
   const siblingDigits = Array.from(siblings).reduce((acc, si) => {
     const sibling = state.cells[si];
@@ -301,4 +271,4 @@ function gameReducer(state: GameState, action: GameAction): GameState {
   throw new Error(`Unhandled action: ${action.type}`);
 }
 
-export { cellSiblings, gameReducer, setup, type GameAction, type GameState };
+export { gameReducer, setup, type GameAction, type GameState };
