--- web-app/src/shared/common.ts
@@ -1,13 +1,0 @@
-import { Digit } from "./digit";
-
-type Cell =
-  | {
-      kind: "given" | "proposed";
-      digit: Digit;
-    }
-  | {
-      kind: "note";
-      digits: Set<Digit>;
-    };
-
-export { type Cell };
